template/launchpad/requirements.txt
boto3
pytz
python-dotenv
streamlit